Refactor the Helper stuff to split out local and remote
