+++ Sources/OAuth/Helper/OAuthHelper.swift
@@ -1,0 +1,4 @@
+protocol OAuthHelper {
+    func assertScopes(_ scopes: [String]?) throws
+    func user() throws -> OAuthUser
+}
